# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для получения информации о человеке (возраст, пол, национальность)",
        "title": "People Information API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/people/search": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Поиск людей по фамилии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фамилия",
                        "name": "last_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xjncx_people-info-api_internal_dto.PersonListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "github_com_xjncx_people-info-api_internal_dto.PersonListResponse": {
            "type": "object",
            "properties": {
                "people": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xjncx_people-info-api_internal_dto.PersonResponse"
                    }
                }
            }
        },
        "github_com_xjncx_people-info-api_internal_dto.PersonResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                }
            }
        }
    }
}